Recast: add dedicated double-precission time type
The library used signed 32bit integers and single precission floating points to represent build times, but on long build times this will overflow. Added a dedicated time type while promoting it to double precission and adjusting all use cases throughout the library.

+++ src/naveditor/include/Editor_Common.h
@@ -53,7 +53,7 @@ protected:
 	float m_lastBuiltTileBmin[3];
 	float m_lastBuiltTileBmax[3];
 
-	float m_totalBuildTimeMs;
+	double m_totalBuildTimeMs;
 
 	bool m_drawActiveTile;
 	bool m_keepInterResults;
@@ -78,7 +78,7 @@ protected:
 	struct FastLZCompressor* m_tcomp;
 	struct MeshProcess* m_tmproc;
 
-	float m_cacheBuildTimeMs;
+	double m_cacheBuildTimeMs;
 	int m_cacheCompressedSize;
 	int m_cacheRawSize;
 	int m_cacheLayerCount;
